Move sbp exposure target modifier into Treatment component
