+++ src/vivarium_nih_us_cvd/components/risks.py
@@ -1,5 +1,3 @@
-from typing import Callable
-
 import pandas as pd
 from vivarium.framework.engine import Builder
 from vivarium.framework.population.manager import PopulationView
@@ -9,23 +7,7 @@ from vivarium_public_health.risks.data_transformations import (
     get_exposure_post_processor,
 )
 
-from vivarium_nih_us_cvd.constants.data_values import (
-    COLUMNS,
-    MEDICATION_ADHERENCE_TYPE,
-    RISK_EXPOSURE_LIMITS,
-    SBP_MEDICATION_LEVEL,
-)
-from vivarium_nih_us_cvd.constants.paths import FILEPATHS
-
-# Format the SBP risk effects file and generate bin edges
-sbp_risk_effects = pd.read_csv(FILEPATHS.SBP_MEDICATION_EFFECTS)
-sbp_risk_effects.loc[
-    sbp_risk_effects["sbp_start_exclusive"].isna(), "sbp_start_exclusive"
-] = -float("inf")
-sbp_risk_effects.loc[
-    sbp_risk_effects["sbp_end_inclusive"].isna(), "sbp_end_inclusive"
-] = float("inf")
-sbp_bin_edges = sorted(set(sbp_risk_effects["sbp_start_exclusive"]).union(set(sbp_risk_effects["sbp_end_inclusive"])))
+from vivarium_nih_us_cvd.constants.data_values import COLUMNS, RISK_EXPOSURE_LIMITS
 
 
 class Risk(Risk_):
@@ -60,8 +42,6 @@ class SBPRisk(Risk_):
     def setup(self, builder: Builder) -> None:
         super().setup(builder)
         self.gbd_exposure = self._get_gbd_exposure_pipeline(builder)
-        self.target_modifier = self._get_target_modifier(builder)
-        self._register_target_modifier(builder)
 
     def _get_gbd_exposure_pipeline(self, builder: Builder) -> Pipeline:
         return builder.value.register_value_producer(
@@ -86,8 +66,6 @@ class SBPRisk(Risk_):
             [
                 self.propensity_column_name,
                 COLUMNS.SBP_MULTIPLIER,
-                COLUMNS.SBP_MEDICATION,
-                COLUMNS.SBP_MEDICATION_ADHERENCE,
             ]
         )
 
@@ -113,54 +91,4 @@ class SBPRisk(Risk_):
 
         return (
             self.gbd_exposure(index) * self.population_view.get(index)[COLUMNS.SBP_MULTIPLIER]
-        )
-
-    def _get_target_modifier(
-        self, builder: Builder
-    ) -> Callable[[pd.Index, pd.Series], pd.Series]:
-        """Apply medication effects"""
-
-        def adjust_target(index: pd.Index, target: pd.Series) -> pd.Series:
-            """Determine the exposure decrease as treatment_efficacy * adherence_score"""
-            pop_view = self.population_view.get(index)
-            mask_adherence = pop_view[COLUMNS.SBP_MEDICATION_ADHERENCE] == MEDICATION_ADHERENCE_TYPE.ADHERENT
-            df_efficacy = pd.DataFrame(
-                {"bin": pd.cut(x=target, bins=sbp_bin_edges, right=True)}
-            )
-            df_efficacy["sbp_start_exclusive"] = df_efficacy["bin"].apply(lambda x: x.left)
-            df_efficacy["sbp_end_inclusive"] = df_efficacy["bin"].apply(lambda x: x.right)
-            df_efficacy = pd.concat([df_efficacy, pop_view[COLUMNS.SBP_MEDICATION]], axis=1)
-            df_efficacy = (
-                df_efficacy.reset_index()
-                .merge(
-                    sbp_risk_effects,
-                    on=["sbp_start_exclusive", "sbp_end_inclusive", COLUMNS.SBP_MEDICATION],
-                    how="left",
-                )
-                .set_index("index")
-            )
-            # Simulants not on treatment mean 0 effect
-            assert set(
-                df_efficacy.loc[df_efficacy["value"].isna(), COLUMNS.SBP_MEDICATION]
-            ) == {SBP_MEDICATION_LEVEL.NO_TREATMENT.DESCRIPTION}
-            df_efficacy.loc[
-                df_efficacy[COLUMNS.SBP_MEDICATION]
-                == SBP_MEDICATION_LEVEL.NO_TREATMENT.DESCRIPTION,
-                "value",
-            ] = 0
-            assert df_efficacy["value"].isna().sum() == 0
-            treatment_efficacy = df_efficacy["value"]
-
-            sbp_decrease = treatment_efficacy * mask_adherence
-
-            return target - sbp_decrease
-
-        return adjust_target
-
-    def _register_target_modifier(self, builder: Builder) -> None:
-        builder.value.register_value_modifier(
-            self.exposure_pipeline_name,
-            modifier=self.target_modifier,
-            # requires_values=[f"{self.risk.name}.exposure"],
-            requires_columns=[COLUMNS.SBP_MEDICATION, COLUMNS.SBP_MEDICATION_ADHERENCE],
         )
