+++ src/vivarium_nih_us_cvd/components/treatment.py
@@ -1,4 +1,4 @@
-from typing import Optional
+from typing import Callable, Optional
 
 import numpy as np
 import pandas as pd
@@ -7,7 +7,7 @@ from vivarium.framework.event import Event
 from vivarium.framework.population import SimulantData
 from vivarium.framework.values import Pipeline
 
-from vivarium_nih_us_cvd.constants import data_values, models
+from vivarium_nih_us_cvd.constants import data_values, models, paths
 from vivarium_nih_us_cvd.utilities import get_random_value_from_normal_distribution
 
 sbp_treatment_map = {
@@ -16,6 +16,21 @@ sbp_treatment_map = {
 ldlc_treatment_map = {
     level.VALUE: level.DESCRIPTION for level in data_values.LDLC_MEDICATION_LEVEL
 }
+
+
+# Format the SBP risk effects file and generate bin edges
+sbp_risk_effects = pd.read_csv(paths.FILEPATHS.SBP_MEDICATION_EFFECTS)
+sbp_risk_effects.loc[
+    sbp_risk_effects["sbp_start_exclusive"].isna(), "sbp_start_exclusive"
+] = -float("inf")
+sbp_risk_effects.loc[
+    sbp_risk_effects["sbp_end_inclusive"].isna(), "sbp_end_inclusive"
+] = float("inf")
+sbp_bin_edges = sorted(
+    set(sbp_risk_effects["sbp_start_exclusive"]).union(
+        set(sbp_risk_effects["sbp_end_inclusive"])
+    )
+)
 
 
 class Treatment:
@@ -38,6 +53,8 @@ class Treatment:
         self.gbd_sbp = builder.value.get_value(data_values.PIPELINES.SBP_GBD_EXPOSURE)
         self.sbp = builder.value.get_value(data_values.PIPELINES.SBP_EXPOSURE)
         self.ldlc = builder.value.get_value(data_values.PIPELINES.LDLC_EXPOSURE)
+        self.target_modifier = self._get_target_modifier(builder)
+        self._register_target_modifier(builder)
 
         columns_created = [
             data_values.COLUMNS.SBP_MEDICATION,
@@ -78,6 +95,70 @@ class Treatment:
             "time_step__cleanup",
             self.on_time_step_cleanup,
             priority=data_values.TIMESTEP_CLEANUP_PRIORITIES.TREATMENT,
+        )
+
+    def _get_target_modifier(
+        self, builder: Builder
+    ) -> Callable[[pd.Index, pd.Series], pd.Series]:
+        """Apply medication effects"""
+
+        def adjust_target(index: pd.Index, target: pd.Series) -> pd.Series:
+            """Determine the exposure decrease as treatment_efficacy * adherence_score"""
+            pop_view = self.population_view.get(index)
+            mask_adherence = (
+                pop_view[data_values.COLUMNS.SBP_MEDICATION_ADHERENCE]
+                == data_values.MEDICATION_ADHERENCE_TYPE.ADHERENT
+            )
+            df_efficacy = pd.DataFrame(
+                {"bin": pd.cut(x=target, bins=sbp_bin_edges, right=True)}
+            )
+            df_efficacy["sbp_start_exclusive"] = df_efficacy["bin"].apply(lambda x: x.left)
+            df_efficacy["sbp_end_inclusive"] = df_efficacy["bin"].apply(lambda x: x.right)
+            df_efficacy = pd.concat(
+                [df_efficacy, pop_view[data_values.COLUMNS.SBP_MEDICATION]], axis=1
+            )
+            df_efficacy = (
+                df_efficacy.reset_index()
+                .merge(
+                    sbp_risk_effects,
+                    on=[
+                        "sbp_start_exclusive",
+                        "sbp_end_inclusive",
+                        data_values.COLUMNS.SBP_MEDICATION,
+                    ],
+                    how="left",
+                )
+                .set_index("index")
+            )
+            # Simulants not on treatment mean 0 effect
+            assert set(
+                df_efficacy.loc[
+                    df_efficacy["value"].isna(), data_values.COLUMNS.SBP_MEDICATION
+                ]
+            ) == {data_values.SBP_MEDICATION_LEVEL.NO_TREATMENT.DESCRIPTION}
+            df_efficacy.loc[
+                df_efficacy[data_values.COLUMNS.SBP_MEDICATION]
+                == data_values.SBP_MEDICATION_LEVEL.NO_TREATMENT.DESCRIPTION,
+                "value",
+            ] = 0
+            assert df_efficacy["value"].isna().sum() == 0
+            treatment_efficacy = df_efficacy["value"]
+
+            sbp_decrease = treatment_efficacy * mask_adherence
+
+            return target - sbp_decrease
+
+        return adjust_target
+
+    def _register_target_modifier(self, builder: Builder) -> None:
+        builder.value.register_value_modifier(
+            data_values.PIPELINES.SBP_EXPOSURE,
+            modifier=self.target_modifier,
+            # requires_values=[f"{self.risk.name}.exposure"],
+            requires_columns=[
+                data_values.COLUMNS.SBP_MEDICATION,
+                data_values.COLUMNS.SBP_MEDICATION_ADHERENCE,
+            ],
         )
 
     def on_initialize_simulants(self, pop_data: SimulantData) -> None:
